refactor(domain): move node files to subfolder
