+++ src/core/domain/node/NodeKind.ts
