+++ src/core/domain/node/NodeProperties.ts
