+++ src/core/domain/node/UIScreenNode.ts
@@ -1,6 +1,6 @@
-import { BoardNode } from './board/BoardNode';
-import type { BoardProjection } from './board/BoardProjection';
-import { GridPosition } from './board/GridPosition';
+import { BoardNode } from '../board/BoardNode';
+import type { BoardProjection } from '../board/BoardProjection';
+import { GridPosition } from '../board/GridPosition';
 
 export class UIScreenNode extends BoardNode {
   constructor(id: string, label: string, position: GridPosition) {
