+++ src/core/store/useBoardStore.ts
@@ -2,14 +2,14 @@ import { create } from 'zustand';
 import { useShallow } from 'zustand/react/shallow';
 import { GridBoard } from '../domain/board/GridBoard';
 import { type BoardProjection } from '../domain/board/BoardProjection';
-import { DomainEventNode } from '../domain/DomainEventNode';
-import { CommandNode } from '../domain/CommandNode';
+import { DomainEventNode } from '../domain/node/DomainEventNode';
+import { CommandNode } from '../domain/node/CommandNode';
 import { type GridBoardRepository } from '../domain/board/GridBoardRepository';
-import { ReadModelNode } from '../domain/ReadModelNode';
-import { PolicyNode } from '../domain/PolicyNode';
-import { UIScreenNode } from '../domain/UIScreenNode';
-import { type NodeKind } from '../domain/NodeKind';
-import { type NodeProperties, createDefaultNodeProperties } from '../domain/NodeProperties';
+import { ReadModelNode } from '../domain/node/ReadModelNode';
+import { PolicyNode } from '../domain/node/PolicyNode';
+import { UIScreenNode } from '../domain/node/UIScreenNode';
+import { type NodeKind } from '../domain/node/NodeKind';
+import { type NodeProperties, createDefaultNodeProperties } from '../domain/node/NodeProperties';
 import { AddDomainEventNodeCommand } from '../usecases/commands/AddDomainEventNode/AddDomainEventNodeCommand';
 import { AddDomainEventNodeCommandHandler } from '../usecases/commands/AddDomainEventNode/AddDomainEventNodeCommandHandler';
 import { MoveNodeCommand } from '../usecases/commands/MoveNode/MoveNodeCommand';
@@ -52,7 +52,7 @@ import { ExportJSONQuery } from '../usecases/queries/ExportJSON/ExportJSONQuery'
 import { ExportJSONQueryHandler } from '../usecases/queries/ExportJSON/ExportJSONQueryHandler';
 import { ExportMarkdownQuery } from '../usecases/queries/ExportMarkdown/ExportMarkdownQuery';
 import { ExportMarkdownQueryHandler } from '../usecases/queries/ExportMarkdown/ExportMarkdownQueryHandler';
-import { type NodeLink } from '../domain/NodeLink';
+import { type NodeLink } from '../domain/node/NodeLink';
 import { type ConnectionType } from '../domain/ConnectionType';
 import { VerticalSliceCollection } from '../domain/vertical-slice/VerticalSliceCollection';
 import { type VerticalSliceRepository } from '../domain/VerticalSliceRepository';
