+++ tests/core/usecases/ExportJSON/ExportJSONQueryHandler.test.ts
@@ -1,14 +1,14 @@
 import { describe, it, expect } from 'vitest';
 import { GridBoard } from '../../../../src/core/domain/board/GridBoard';
-import { DomainEventNode } from '../../../../src/core/domain/DomainEventNode';
-import { CommandNode } from '../../../../src/core/domain/CommandNode';
-import { ReadModelNode } from '../../../../src/core/domain/ReadModelNode';
-import { PolicyNode } from '../../../../src/core/domain/PolicyNode';
-import { UIScreenNode } from '../../../../src/core/domain/UIScreenNode';
+import { DomainEventNode } from '../../../../src/core/domain/node/DomainEventNode';
+import { CommandNode } from '../../../../src/core/domain/node/CommandNode';
+import { ReadModelNode } from '../../../../src/core/domain/node/ReadModelNode';
+import { PolicyNode } from '../../../../src/core/domain/node/PolicyNode';
+import { UIScreenNode } from '../../../../src/core/domain/node/UIScreenNode';
 import { VerticalSliceCollection } from '../../../../src/core/domain/vertical-slice/VerticalSliceCollection';
 import { VerticalSlice } from '../../../../src/core/domain/vertical-slice/VerticalSlice';
 import { Scenario } from '../../../../src/core/domain/Scenario';
-import { type NodeProperties } from '../../../../src/core/domain/NodeProperties';
+import { type NodeProperties } from '../../../../src/core/domain/node/NodeProperties';
 import { ExportJSONQuery } from '../../../../src/core/usecases/queries/ExportJSON/ExportJSONQuery';
 import { ExportJSONQueryHandler } from '../../../../src/core/usecases/queries/ExportJSON/ExportJSONQueryHandler';
